Function media_type::push_params_to_buffer

pub fn push_params_to_buffer<S, I, IN, IV>(
    buffer: &mut String, 
    params: I
) -> Result<Vec<ParamIndices>, Error> where
    S: Spec,
    I: IntoIterator<Item = (IN, IV)>,
    IN: AsRef<str>,
    IV: AsRef<str>, 

Push all parameters formatted to the output buffer

each parameter is preceded by "; " followed by then "=", then . If the value needed to be quoted it will be quoted, if the value needs to be encoded it's encoded (and "*" is added to the parameter name.

Error

an error is returned if a parameter name is not valid for the given Spec S.