Trait media_type::AsciiCaseInsensitiveEq

pub trait AsciiCaseInsensitiveEq<Rhs> where
    Rhs: ?Sized, {
    fn eq_ignore_ascii_case(&self, other: &Rhs) -> bool;
}

Analogous to PartialEq, but with ascii case insensitive equality

Required methods

fn eq_ignore_ascii_case(&self, other: &Rhs) -> bool

compares this instance with other with a ascii case insensitive comparsion algorithm

Note that this is ascii case insensitivity. Which means this will not work well/as expected if the content contain non ascii characters. E.g. the upercase of "ß" was "SS" but by know there is also a "ẞ" used in all caps writing.

Loading content...